--- mia/helpers/__init__.py
@@ -1,3 +1,0 @@
-"""
-This sub-module contains the helpers for "mia".
-"""
